refactor(p2): extract SubAgentTaskStepView into sibling (stage 4 of 4, final)
Moves the SubAgent UI system from ChatMessageTools.kt L1645-L1951
(parent ae597a0d) into a new sibling ChatMessageSubAgentStep.kt.
Bodies are byte-equal with the parent commit.

In the stage-3 coordinator (2dd446a1) the same block had shifted to
L866-L1180; stripped L865-L1180 (leading blank + block) so coordinator
EOF is `}\n` (single newline). Coordinator final state: 838 lines (down
from parent's 2605 — ≈ 32% of original size).

Stage 4 is the smallest cut by line count and the cleanest: ZERO new
widenings in either coord or sibling. All required visibility upgrades
were already done by prior stages:
  - getStringContent / ToolNames / MemoryActions widened in stage 1
  - AgentToolStatus / AgentToolKind / AgentToolCallCapsule /
    PHASE_LABEL_INTERVAL_MS widened in stage 3
  - payloadObject / extractFinalSubAgentText /
    extractSubAgentDisplayTextFromTranscript widened in stage 3 Council
    sibling

The `internal const val PHASE_LABEL_INTERVAL_MS = 5_000L` modifier carries
over to the new SubAgent sibling byte-equal, so Council sibling continues
to see it through the same-package cross-sibling reference.

Five `by` delegations in extracted body (showSheet, phaseIndex, liveText,
transcriptText, followBottom) — sibling imports
androidx.compose.runtime.getValue + setValue explicitly per playbook §5.3.

Coordinator strip: 1180 → 864 (-316 body+blank) → 838 (-26 import cleanup).
Imports trimmed: 26 this-refactor orphans (LinearEasing, awaitEachGesture,
awaitFirstDown, Spacer, fillMaxHeight, LazyColumn, rememberScrollState,
SelectionContainer, verticalScroll, HorizontalDivider, ModalBottomSheet,
rememberModalBottomSheetState, LaunchedEffect, collectAsState,
mutableIntStateOf, snapshotFlow, pointerInput, LocalContext, delay,
MutableStateFlow, JsonPrimitive, SubAgentDefinitions, SubAgentManager,
SubAgentRunStatus, MarkdownBlock, File). Several (JsonPrimitive,
LaunchedEffect) had refs partly in stage 1/2/3 ranges but the import
stayed live until SubAgent's last reference left coord at this stage.

Pre-existing dead imports preserved (CLAUDE.md §3): Box, BoxWithConstraints,
CircleShape, FilledTonalButton, FilterChip, LocalContentColor, Color,
getValue, setValue (last two used implicitly via `by` — 3 sites in final
coord).

## Final cut tally

| Stage | Block | Sibling | Lines | Coord widenings | Sibling widenings |
|---|---|---|---|---|---|
| 1 | ToolPreviewSheets | ChatMessageToolPreviewSheet.kt | 336 | 3 (getStringContent, ToolNames, MemoryActions) | 1 (ToolCallPreviewSheet) |
| 2 | AskUser | ChatMessageAskUserStep.kt | 396 | 0 | 2 (AskUserToolStep, ToolDenyReasonDialog) |
| 3 | Council | ChatMessageCouncilStep.kt | 654 | 4 (AgentToolStatus, AgentToolKind, AgentToolCallCapsule, PHASE_LABEL_INTERVAL_MS) | 3 (payloadObject, extractFinalSubAgentText, extractSubAgentDisplayTextFromTranscript) |
| 4 | SubAgent | ChatMessageSubAgentStep.kt | 315 | 0 | 0 |

Coord 2605 → 838 lines. Stage 3's `extractSubAgentDisplayTextFromTranscript`
widening was caught at compile gate (pre-cut §3.2 missed it; addressed in
stage 3 commit body).

## Verification (final stage)

- :app:compileDebugKotlin: BUILD SUCCESSFUL
- :app:testDebugUnitTest: 467 tests / 3 failed (baseline preserved)
- :app:assembleDebug / :app:assembleNotion / :app:assembleRefactortest:
  all BUILD SUCCESSFUL, APKs produced at app/build/outputs/apk/{debug,notion,
  refactortest}/*.apk

Baseline 467 tests / 3 failed preserved across all 4 stages.
